Refactoring to use pathlib library and modules
